+++ sketch_oxide/src/frequency/mod.rs
@@ -17,6 +17,7 @@ mod space_saving;
 mod space_saving_pm;
 mod spread_sketch;
 mod tower_sketch;
+mod unbiased_space_saving;
 mod waving_sketch;
 
 pub use conservative_count_min::ConservativeCountMin;
@@ -36,4 +37,5 @@ pub use space_saving::SpaceSaving;
 pub use space_saving_pm::SpaceSavingPlusMinus;
 pub use spread_sketch::SpreadSketch;
 pub use tower_sketch::TowerSketch;
+pub use unbiased_space_saving::UnbiasedSpaceSaving;
 pub use waving_sketch::WavingSketch;